+++ tradingbot/strategies/ema_strategy.py
@@ -1,5 +1,5 @@
 """EMA Strategy."""
-from .strategy import Strategy
+from tradingbot.strategies.strategy import Strategy
 from typing import Union
 from tradingbot.order import (Order,
                               MutableOrderDetails,
@@ -7,10 +7,10 @@ from tradingbot.order import (Order,
                               OrderPrice,
                               OrderType)
 
-from tradingbot.trading_methods import EMA, get_pivots
+from tradingbot.trading_methods import EMA, get_pivots, get_pip
 from tradingbot.ohlc import OHLC
 from tradingbot.utils import create_magic_number
-from tradingbot.forex_client import MT_Client
+from datetime import datetime
 
 
 class EMA_strategy(Strategy):
@@ -24,8 +24,12 @@ class EMA_strategy(Strategy):
       self,
       ohlc: OHLC,
       symbol: str,
+      now_date: datetime
   ) -> Union[Order, None]:
     """Return an order if the strategy is triggered."""
+    # This strategy does not use the date
+    _ = now_date
+
     # Calculate the EMAs
     ema_20 = EMA(ohlc.close, 20)
     ema_50 = EMA(ohlc.close, 50)
@@ -34,7 +38,6 @@ class EMA_strategy(Strategy):
     p = get_pivots(ohlc.high, left=6, right=3, n_pivot=2, max_min='max')
     max_pivot_1 = p[0][0]  # most recent
     max_pivot_2 = p[1][0]
-
     p = get_pivots(ohlc.low, left=6, right=3, n_pivot=2, max_min='min')
     min_pivot_1 = p[0][0]  # most recent
     min_pivot_2 = p[1][0]
@@ -46,7 +49,7 @@ class EMA_strategy(Strategy):
 
     # Obtain common features
     magic = create_magic_number()
-    pip = MT_Client.get_pip(symbol)
+    pip = get_pip(symbol)
 
     if ema_20[-1] >= ema_50[-1] and upper_tendency:
       return Order(
@@ -56,10 +59,8 @@ class EMA_strategy(Strategy):
                   stop_loss=ema_50[-1] - 10 * pip
               )),
           ImmutableOrderDetails(
-              symbol=symbol,
-              order_type=OrderType.BUY,
-              magic=magic,
-              comment=self.strategy_name)
+              symbol=symbol, order_type=OrderType.BUY,
+              magic=magic, comment=self.strategy_name)
       )
     elif lower_tendency:
       return Order(
@@ -69,10 +70,8 @@ class EMA_strategy(Strategy):
                   stop_loss=ema_50[-1] + 10 * pip
               )),
           ImmutableOrderDetails(
-              symbol=symbol,
-              order_type=OrderType.SELL,
-              magic=magic,
-              comment=self.strategy_name)
+              symbol=symbol, order_type=OrderType.SELL,
+              magic=magic, comment=self.strategy_name)
       )
 
     return None